ChatWebApp/requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.0.6
django-channels==0.7.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==13.15.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Markdown==3.4.1
msgpack==1.0.4
oauthlib==3.2.0
Pillow==9.2.0
pip==22.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
setuptools==61.2.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
wheel==0.37.1
wincertstore==0.2
zope.interface==5.4.0
